+++ packages/discovery-react-components/src/components/DiscoverySearch/__tests__/DiscoverySearch.test.tsx
@@ -7,7 +7,7 @@ import {
   SearchContext
 } from '../DiscoverySearch';
 import DiscoveryV2 from 'ibm-watson/discovery/v2';
-import { createDummyResponsePromise } from '../../../utils/testingUtils';
+import { createDummyResponsePromise } from 'utils/testingUtils';
 interface Setup {
   fullTree: JSX.Element;
   result: RenderResult;