+++ packages/discovery-react-components/src/components/DocumentPreview/components/Highlight/TableHighlight.tsx
@@ -1,11 +1,11 @@
 import { FC, useEffect, useState, ReactElement } from 'react';
 import get from 'lodash/get';
 import { QueryResult, QueryTableResult } from 'ibm-watson/discovery/v2';
-import processDoc, { ProcessedDoc, Table } from '../../../../utils/document/processDoc';
-import { getTextMappings } from '../../utils/documentData';
+import { processDoc, ProcessedDoc, Table } from 'utils/document';
+import { getTextMappings } from 'components/DocumentPreview/utils/documentData';
 import { DEFAULT_WIDTH, DEFAULT_HEIGHT } from './constants';
-import { Page } from '../../types';
-import { spansIntersect } from '../../../../utils/document/documentUtils';
+import { Page } from 'components/DocumentPreview/types';
+import { spansIntersect } from 'utils/document/documentUtils';
 
 interface Props {
   /**