+++ packages/discovery-react-components/src/components/SearchFacets/components/FacetsGroups/MultiSelectFacetsGroup.tsx
@@ -1,12 +1,12 @@
 import React, { FC, useContext, SyntheticEvent } from 'react';
-import { optionClass, optionLabelClass } from './facetGroupClasses';
+import { optionClass, optionLabelClass } from '../../cssClasses';
 import { Checkbox as CarbonCheckbox } from 'carbon-components-react';
-import { SearchContext } from '../../../DiscoverySearch/DiscoverySearch';
+import { SearchContext } from 'components/DiscoverySearch/DiscoverySearch';
 import {
   SelectableDynamicFacets,
   SelectableQueryTermAggregationResult,
   InternalQueryTermAggregation
-} from '../../utils/searchFacetInterfaces';
+} from 'components/SearchFacets/utils/searchFacetInterfaces';
 import get from 'lodash/get';
 
 interface MultiSelectFacetsGroupProps {