+++ packages/discovery-react-components/src/components/SearchResults/components/Result/__tests__/Result.test.tsx
@@ -6,11 +6,11 @@ import {
   SearchContextIFC,
   searchResponseStoreDefaults,
   fetchDocumentsResponseStoreDefaults
-} from '../../../../DiscoverySearch/DiscoverySearch';
+} from 'components/DiscoverySearch/DiscoverySearch';
 import DiscoveryV2 from 'ibm-watson/discovery/v2';
 
-import { wrapWithContext, browserWindow } from '../../../../../utils/testingUtils';
-import { SearchResults } from '../../../SearchResults';
+import { wrapWithContext, browserWindow } from 'utils/testingUtils';
+import { SearchResults } from 'components/SearchResults/SearchResults';
 
 describe('<Result />', () => {
   let context: Partial<SearchContextIFC>;