+++ packages/discovery-react-components/src/utils/document/processDoc.ts
@@ -4,7 +4,7 @@ import { QueryResult } from 'ibm-watson/discovery/v2';
 import { isRelationObject } from './nonContractUtils';
 import { getId } from './idUtils';
 import transformEnrichment from './transformEnrichment';
-import { getEnrichmentName } from '../../components/CIDocument/utils/enrichmentUtils';
+import { getEnrichmentName } from 'components/CIDocument/utils/enrichmentUtils';
 import { spansIntersect } from './documentUtils';
 
 // split HTML into "sections" based on these top level tag(s)
@@ -75,7 +75,7 @@ export interface Table {
  * @param {Boolean} options.itemMap return item mapping into 'sections'
  * @throws {ParsingError}
  */
-export default async function processDoc(
+export async function processDoc(
   // eslint-disable-next-line @typescript-eslint/camelcase
   { html, enriched_html }: QueryResult,
   options?: Options