+++ examples/discovery-search-app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.2-beta.2 (2019-12-10)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## 1.0.2-beta.1 (2019-12-10)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
+## 1.0.2-beta.0 (2019-12-09)
+
+**Note:** Version bump only for package discovery-search-app
+
+
+
+
+
 ## 1.0.1 (2019-12-06)
 
 **Note:** Version bump only for package discovery-search-app